+++ Action/ComboMeanPlayerBase.h
@@ -19,7 +19,7 @@ public:
 	@param &_playerCanNotMoveTime 攻撃後プレイヤーの行動不可時間、硬直を所持する変数への参照
 	@param &_waitTimeForNextAttack 攻撃後硬直から次の攻撃までの時間を所持する変数への参照
 	*/
-	void Attack(const Vector3& _pos, const int& _direction, float& _playerCanNotMoveTime, float& _waitTimeForNextAttack);
+	void Attack(const Vector3& _pos, const float& _direction, float& _playerCanNotMoveTime, float& _waitTimeForNextAttack);
 	/*
 	@fn AttackPlayerComponentから呼ばれる。攻撃状態をリセットする。
 	@detail AttackPlayerComponent内で管理されている、次の攻撃までの時間waitTimeForNextAttackがゼロになったときに呼ばれる。
