+++ Action/ThrowWeapon.cpp
@@ -4,7 +4,7 @@
 #include "ColliderComponent.h"
 #include "RotateComponent.h"
 
-ThrowWeapon::ThrowWeapon(const Vector3& _pos, const int& _direction, const int& _waitTime, const Tag& _tag) :
+ThrowWeapon::ThrowWeapon(const Vector3& _pos, const float& _direction, const int& _waitTime, const Tag& _tag) :
 	WeaponBase(_pos, _waitTime, 150, _direction, _tag, 1, Vector3(100, 30, 5),0.4f, "Assets/Model/Weapon/SK_Dual_Blade_Arrow.gpmesh", "Assets/Model/Weapon/SK_Dual_Blade_Arrow.gpskel"),
 	velocity(Vector3::Zero),
 	collided(false)
@@ -47,7 +47,7 @@ void ThrowWeapon::UpdateWeaponGameObject(float _deltaTime)
 	{
 		if (!slow)
 		{
-			velocity.y -= (27.0 * _deltaTime);
+			velocity.y -= (27.0f * _deltaTime);
 		}
 		SetPosition(position + velocity);
 	}
